Use array of schema files in readme example
Make it more clear in the documentation that multiple schema files can
be provided as an array. Single schema file string input is still
supported too, but the array version is more general so if we have to
choose one for the documentation then it's better to show that one.

Fixes #2.