Pass --noout if normalization option isn't set
Don't output anything through the normalization property of the return
value if the params didn't explicitly set normalization option. Xmllint
default is to print back the original input XML, which seems rather
useless. It can also come with a nontrivial memory & perf costs when the
input XML is large: the data must be sent from wasm to worker to the
host Node program.