Fix worker import path in browsers
In browsers, new Worker(relativePath) is relative to entry scripts base
path, not the current file's path. That won't work for us, there's no
way to know where the script is located relative to the library
consumers base path.

Instead, use the URL constructor that will create an absolute import url
path based on the basename that we provide as second argument. We get
the basename from import.meta.url.

This should work fine even if the library consumer is using a bundler
instead of native ES6 modules. At least Webpack 5 and Parcel work with
this pattern.

I also unified the previously separated path to worker file and its
construction to one statement, because that's how it's done in all of
the Webpack examples. Maybe it would've been able to pick it up from a
separate variable, maybe not. This way feels safer anyway.