Add demo for usage with a bundler
xmllint-wasm should work fine with a bundler like Parcel that supports
WebWorker loading with the new URL(.., import.meta.url) syntax. Add a
script that can be used to verify that the bundler support is still
working.

NOTE: I had to move the script loading because down to body because
there seems to be a bug(?) in Parcel that it tries to evaluate the
script tag with type=module too early. Spec says that module inline
scripts should be automatically deferred and thus the document should be
ready by then, but Parcel does not seem to respect that, causing the
document.querySelector(..).addEventListener line to error. Moving at the
bottom of body fixed it.